+++ scripts/editor/index.js
@@ -1,45 +1,44 @@
 // All exports are sorted in alphabetical order.
 
 export {
-	overrideInnerBlockAttributes,
-	overrideInnerBlockSimpleWrapperAttributes,
-	checkAttr,
-	checkAttrResponsive,
-	getAttrKey,
-	props,
-} from './attributes';
-export { getPaletteColors } from './colors';
+  overrideInnerBlockAttributes,
+  overrideInnerBlockSimpleWrapperAttributes,
+  checkAttr,
+  checkAttrResponsive,
+  getAttrKey,
+  props,
+} from "./attributes";
+export { getPaletteColors } from "./colors";
 export {
-	outputCssVariablesGlobal,
-	outputCssVariables,
-	hexToRgb,
-	getUnique,
-} from './css-variables';
+  outputCssVariablesGlobal,
+  outputCssVariables,
+  hexToRgb,
+  getUnique,
+} from "./css-variables";
 export {
-	GutenbergBlock,
-	lockIfUndefined,
-	lockPostEditing,
-	unlockPostEditing,
-	getResponsiveLegacyData,
-	generateOptionsFromValue,
-} from './editor';
-export { getFetchWpApi, fetchFromWpRest, wpSearchRoute } from './fetch';
-export { inserter } from './inserter';
-export { getOption, getOptionColors, getOptions } from './options';
-export { pasteInto } from './paste-handler';
+  GutenbergBlock,
+  lockIfUndefined,
+  lockPostEditing,
+  unlockPostEditing,
+  getResponsiveLegacyData,
+  generateOptionsFromValue,
+} from "./editor";
+export { fetchFromWpRest, fetchFromWpRest, buildWpRestUrl } from "./fetch";
+export { inserter } from "./inserter";
+export { getOption, getOptionColors, getOptions } from "./options";
 export {
-	getAttributes,
-	getExample,
-	getFullBlockName,
-	getFullBlockNameVariation,
-	registerBlocks,
-	registerVariations,
-} from './registration';
-export { bem, selector, responsiveSelectors } from './selectors';
+  getAttributes,
+  getExample,
+  getFullBlockName,
+  getFullBlockNameVariation,
+  registerBlocks,
+  registerVariations,
+} from "./registration";
+export { bem, selector, responsiveSelectors } from "./selectors";
 export {
-	STORE_NAME,
-	BUILD_VERSION,
-	setStore,
-	setStoreGlobalWindow,
-	setConfigFlags,
-} from './store';
+  STORE_NAME,
+  BUILD_VERSION,
+  setStore,
+  setStoreGlobalWindow,
+  setConfigFlags,
+} from "./store";