+++ scripts/helpers/index.js
@@ -1,6 +1,6 @@
 // All exports are sorted in alphabetical order.
 
-export { getDefaultBreakpointNames } from './breakpoints';
-export { cookies } from './cookies';
-export { dynamicImport } from './dynamic-import';
-export { luminanceFromHex, luminanceFromRgb } from './color-helpers';
+export { getDefaultBreakpointNames } from "./breakpoints";
+export { cookies } from "./cookies";
+export { dynamicImport } from "./dynamic-import";
+export { luminanceFromHex, luminanceFromRgb } from "./color-helpers";